Update EventServiceProvider.php - register new created event Update User.php - add logic for checking if image is  not empty Update UsersController.php - dispatch new created event Add event UserRegistered Add listener

+++ app/Models/User.php
@@ -69,7 +69,9 @@ class User extends Authenticatable
      */
     public function setImageAttribute($value)
     {
-        File::delete($this->attributes['image']);
+        if (!empty($this->attributes['image'])) {
+            File::delete($this->attributes['image']);
+        }
         $this->attributes['image'] = $value instanceof UploadedFile ? Upload::process($value) : $value;
     }
